refactor(pyopnsense): split into multiple files
