+++ custom_components/opnsense/pyopnsense/exceptions.py
@@ -1,0 +1,9 @@
+"""Custom exceptions for pyopnsense."""
+
+
+class VoucherServerError(Exception):
+    """Error from Voucher Server."""
+
+
+class UnknownFirmware(Exception):
+    """Unknown current firmware version."""
